+++ src/Loader/Http/HttpLoader.php
@@ -2,26 +2,18 @@
 
 namespace Crwlr\Crawler\Loader\Http;
 
-use Crwlr\Crawler\Loader\Http\Cookies\CookieJar;
 use Crwlr\Crawler\Loader\Http\Exceptions\LoadingException;
 use Crwlr\Crawler\Loader\Http\Messages\RespondedRequest;
 use Crwlr\Crawler\Loader\Http\Politeness\RetryErrorResponseHandler;
-use Crwlr\Crawler\Loader\Http\Politeness\RobotsTxtHandler;
 use Crwlr\Crawler\Loader\Http\Politeness\Throttler;
-use Crwlr\Crawler\Loader\Loader;
 use Crwlr\Crawler\Steps\Filters\FilterInterface;
 use Crwlr\Crawler\UserAgents\UserAgentInterface;
-use Crwlr\Crawler\Utils\RequestKey;
-use Crwlr\Url\Exceptions\InvalidUrlException;
 use Crwlr\Url\Url;
-use Error;
 use Exception;
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\GuzzleException;
-use GuzzleHttp\Psr7\Request;
 use GuzzleHttp\Psr7\Response;
 use HeadlessChromium\Browser;
-use HeadlessChromium\BrowserFactory;
 use HeadlessChromium\Exception\CommunicationException;
 use HeadlessChromium\Exception\NavigationExpired;
 use HeadlessChromium\Exception\NoResponseAvailable;
@@ -31,36 +23,16 @@ use Psr\Http\Client\ClientExceptionInterface;
 use Psr\Http\Client\ClientInterface;
 use Psr\Http\Message\RequestInterface;
 use Psr\Http\Message\ResponseInterface;
-use Psr\Http\Message\UriInterface;
 use Psr\Log\LoggerInterface;
 use Throwable;
 
-class HttpLoader extends Loader
+class HttpLoader extends HttpBaseLoader
 {
     protected ClientInterface $httpClient;
 
-    protected CookieJar $cookieJar;
-
-    protected bool $useCookies = true;
+    protected ?HeadlessBrowserLoaderHelper $browserHelper = null;
 
     protected bool $useHeadlessBrowser = false;
-
-    protected ?string $chromeExecutable = null;
-
-    /**
-     * @var mixed[]
-     */
-    protected array $headlessBrowserOptions = [
-        'windowSize' => [1920, 1000],
-    ];
-
-    protected bool $headlessBrowserOptionsDirty = false;
-
-    protected ?Browser $headlessBrowser = null;
-
-    protected RobotsTxtHandler $robotsTxtHandler;
-
-    protected Throttler $throttler;
 
     /**
      * @var mixed[]
@@ -69,12 +41,6 @@ class HttpLoader extends Loader
         'connect_timeout' => 10,
         'timeout' => 60,
     ];
-
-    protected int $maxRedirects = 10;
-
-    protected bool $retryCachedErrorResponses = false;
-
-    protected bool $writeOnlyCache = false;
 
     /**
      * @var array<int, FilterInterface>
@@ -91,37 +57,14 @@ class HttpLoader extends Loader
         ?ClientInterface $httpClient = null,
         ?LoggerInterface $logger = null,
         ?Throttler $throttler = null,
-        protected RetryErrorResponseHandler $retryErrorResponseHandler = new RetryErrorResponseHandler(),
+        RetryErrorResponseHandler $retryErrorResponseHandler = new RetryErrorResponseHandler(),
         array $defaultGuzzleClientConfig = [],
     ) {
-        parent::__construct($userAgent, $logger);
-
-        $this->retryErrorResponseHandler->setLogger($this->logger);
+        parent::__construct($userAgent, $logger, $throttler, $retryErrorResponseHandler);
 
         $this->httpClient = $httpClient ?? new Client($this->mergeClientConfigWithDefaults($defaultGuzzleClientConfig));
 
-        $this->onSuccess(function (RequestInterface $request, ResponseInterface $response, LoggerInterface $logger) {
-            $logger->info('Loaded ' . $request->getUri()->__toString());
-        });
 
-        $this->onError(function (RequestInterface $request, Exception|Error|ResponseInterface $exceptionOrResponse, $logger) {
-            $logMessage = 'Failed to load ' . $request->getUri()->__toString() . ': ';
-
-            if ($exceptionOrResponse instanceof ResponseInterface) {
-                $logMessage .= 'got response ' . $exceptionOrResponse->getStatusCode() . ' - ' .
-                    $exceptionOrResponse->getReasonPhrase();
-            } else {
-                $logMessage .= $exceptionOrResponse->getMessage();
-            }
-
-            $logger->error($logMessage);
-        });
-
-        $this->cookieJar = new CookieJar();
-
-        $this->robotsTxtHandler = new RobotsTxtHandler($this, $this->logger);
-
-        $this->throttler = $throttler ?? new Throttler();
     }
 
     /**
@@ -223,21 +166,18 @@ class HttpLoader extends Loader
         }
     }
 
-    public function dontUseCookies(): static
+    public function useHeadlessBrowser(): static
     {
-        $this->useCookies = false;
+        $this->useHeadlessBrowser = true;
 
         return $this;
     }
 
-    public function flushCookies(): void
+    public function useHttpClient(): static
     {
-        $this->cookieJar->flush();
-    }
+        $this->useHeadlessBrowser = false;
 
-    public function useHeadlessBrowser(): static
-    {
-        $this->useHeadlessBrowser = true;
+        $this->browserHelper()->closeBrowser();
 
         return $this;
     }
@@ -247,105 +187,31 @@ class HttpLoader extends Loader
         return $this->useHeadlessBrowser;
     }
 
-    public function useHttpClient(): static
-    {
-        $this->useHeadlessBrowser = false;
-
-        $this->headlessBrowser = null;
-
-        return $this;
-    }
-
     /**
-     * @param mixed[] $options
+     * @param array<string, mixed> $options
      */
     public function setHeadlessBrowserOptions(array $options): static
     {
-        $this->headlessBrowserOptions = $options;
-
-        $this->headlessBrowserOptionsDirty = true;
+        $this->browserHelper()->setOptions($options);
 
         return $this;
     }
 
     /**
-     * @param mixed[] $options
+     * @param array<string, mixed> $options
      */
     public function addHeadlessBrowserOptions(array $options): static
     {
-        foreach ($options as $key => $value) {
-            $this->headlessBrowserOptions[$key] = $value;
-        }
-
-        $this->headlessBrowserOptionsDirty = true;
+        $this->browserHelper()->addOptions($options);
 
         return $this;
     }
 
     public function setChromeExecutable(string $executable): static
     {
-        $this->chromeExecutable = $executable;
+        $this->browserHelper()->setExecutable($executable);
 
         return $this;
-    }
-
-    public function setMaxRedirects(int $maxRedirects): static
-    {
-        $this->maxRedirects = $maxRedirects;
-
-        return $this;
-    }
-
-    public function robotsTxt(): RobotsTxtHandler
-    {
-        return $this->robotsTxtHandler;
-    }
-
-    public function throttle(): Throttler
-    {
-        return $this->throttler;
-    }
-
-    public function retryCachedErrorResponses(): static
-    {
-        $this->retryCachedErrorResponses = true;
-
-        return $this;
-    }
-
-    public function writeOnlyCache(): static
-    {
-        $this->writeOnlyCache = true;
-
-        return $this;
-    }
-
-    public function cacheOnlyWhereUrl(FilterInterface $filter): static
-    {
-        $this->cacheUrlFilters[] = $filter;
-
-        return $this;
-    }
-
-    /**
-     * @throws Exception
-     */
-    public function useProxy(string $proxyUrl): void
-    {
-        $this->checkIfProxiesCanBeUsed();
-
-        $this->proxies = new ProxyManager([$proxyUrl]);
-    }
-
-    /**
-     * @param string[] $proxyUrls
-     * @throws Exception
-     */
-    public function useRotatingProxies(array $proxyUrls): void
-    {
-        $this->checkIfProxiesCanBeUsed();
-
-        $this->proxies = new ProxyManager($proxyUrls);
     }
 
     /**
@@ -375,120 +241,6 @@ class HttpLoader extends Loader
         }
 
         return $merged;
-    }
-
-    /**
-     * @throws LoadingException
-     * @throws Exception
-     */
-    protected function isAllowedToBeLoaded(UriInterface $uri, bool $throwsException = false): bool
-    {
-        if (!$this->robotsTxtHandler->isAllowed($uri)) {
-            $message = 'Crawler is not allowed to load ' . $uri . ' according to robots.txt file.';
-
-            $this->logger->warning($message);
-
-            if ($throwsException) {
-                throw new LoadingException($message);
-            }
-
-            return false;
-        }
-
-        return true;
-    }
-
-    /**
-     * @throws \Psr\SimpleCache\InvalidArgumentException
-     * @throws Exception
-     */
-    protected function getFromCache(RequestInterface $request): ?RespondedRequest
-    {
-        if (!$this->cache || $this->writeOnlyCache) {
-            return null;
-        }
-
-        $key = RequestKey::from($request);
-
-        if ($this->cache->has($key)) {
-            $this->logger->info('Found ' . $request->getUri()->__toString() . ' in cache.');
-
-            $respondedRequest = $this->cache->get($key);
-
-            // Previously, until v0.7 just used serialized arrays. Leave this for backwards compatibility.
-            if (is_array($respondedRequest)) {
-                $respondedRequest = RespondedRequest::fromArray($respondedRequest);
-            }
-
-            if ($this->retryCachedErrorResponses && $respondedRequest->response->getStatusCode() >= 400) {
-                $this->logger->info('Cached response was an error response, retry.');
-
-                return null;
-            }
-
-            return $respondedRequest;
-        }
-
-        return null;
-    }
-
-    /**
-     * @throws \Psr\SimpleCache\InvalidArgumentException
-     */
-    protected function addToCache(RespondedRequest $respondedRequest): void
-    {
-        if ($this->cache && $this->shouldResponseBeCached($respondedRequest)) {
-            $this->cache->set($respondedRequest->cacheKey(), $respondedRequest);
-        }
-    }
-
-    protected function shouldResponseBeCached(RespondedRequest $respondedRequest): bool
-    {
-        if (!empty($this->cacheUrlFilters)) {
-            foreach ($this->cacheUrlFilters as $filter) {
-                foreach ($respondedRequest->allUris() as $url) {
-                    if (!$filter->evaluate($url)) {
-                        return false;
-                    }
-                }
-            }
-        }
-
-        return true;
-    }
-
-    /**
-     * @throws InvalidArgumentException
-     */
-    protected function validateSubjectType(RequestInterface|string $requestOrUri): RequestInterface
-    {
-        if (is_string($requestOrUri)) {
-            try {
-                return new Request('GET', Url::parsePsr7($requestOrUri));
-            } catch (InvalidUrlException) {
-                throw new InvalidArgumentException('Invalid URL.');
-            }
-        }
-
-        return $requestOrUri;
-    }
-
-    /**
-     * @throws Exception
-     */
-    protected function prepareRequest(RequestInterface $request): RequestInterface
-    {
-        $request = $request->withHeader('User-Agent', $this->userAgent->__toString());
-
-        // When writing tests I found that guzzle somehow messed up headers with multiple strings as value in the PSR-7
-        // request object. It sent only the last part of the array, instead of concatenating the array of strings to a
-        // comma separated string. Don't know if that happens with all handlers (curl, stream), will investigate
-        // further. But until this is fixed, we just prepare the headers ourselves.
-        foreach ($request->getHeaders() as $headerName => $headerValues) {
-            $request = $request->withHeader($headerName, $request->getHeaderLine($headerName));
-        }
-
-        return $this->addCookiesToRequest($request);
     }
 
     /**
@@ -636,7 +388,7 @@ class HttpLoader extends Loader
             function ($params) use (&$statusCode, &$responseHeaders) {
                 $statusCode = $params['response']['status'];
 
-                $responseHeaders = $this->sanitizeResponseHeaders($params['response']['headers']);
+                $responseHeaders = $this->browserHelper()->sanitizeResponseHeaders($params['response']['headers']);
             }
         );
 
@@ -656,107 +408,24 @@ class HttpLoader extends Loader
         );
     }
 
+    protected function browserHelper(): HeadlessBrowserLoaderHelper
+    {
+        if (!$this->browserHelper) {
+            $this->browserHelper = new HeadlessBrowserLoaderHelper();
+        }
+
+        return $this->browserHelper;
+    }
+
     /**
      * @throws Exception
      */
     protected function getBrowser(RequestInterface $request): Browser
     {
-        if (!$this->headlessBrowser || $this->shouldRenewHeadlessBrowserInstance()) {
-            $this->headlessBrowser?->close();
-
-            $options = $this->headlessBrowserOptions;
-
-            $options['userAgent'] = $this->userAgent->__toString();
-
-            $options['headers'] = array_merge(
-                $options['headers'] ?? [],
-                $this->prepareRequestHeadersForHeadlessBrowser($request->getHeaders()),
-            );
-
-            if (!empty($this->proxies)) {
-                $options['proxyServer'] = $this->proxies->getProxy();
-            }
-
-            $this->headlessBrowser = (new BrowserFactory($this->chromeExecutable))->createBrowser($options);
-
-            $this->headlessBrowserOptionsDirty = false;
+        if (!empty($this->proxies)) {
+            return $this->browserHelper()->getBrowser($request, $this->proxies->getProxy());
         }
 
-        return $this->headlessBrowser;
-    }
-
-    protected function shouldRenewHeadlessBrowserInstance(): bool
-    {
-        return $this->headlessBrowserOptionsDirty || ($this->proxies && $this->proxies->hasMultipleProxies());
-    }
-
-    protected function addCookiesToJar(RespondedRequest $respondedRequest): void
-    {
-        if ($this->useCookies) {
-            try {
-                $this->cookieJar->addFrom($respondedRequest->effectiveUri(), $respondedRequest->response);
-            } catch (Exception $exception) {
-                $this->logger->warning('Problem when adding cookies to the Jar: ' . $exception->getMessage());
-            }
-        }
-    }
-
-    /**
-     * @throws Exception
-     */
-    protected function addCookiesToRequest(RequestInterface $request): RequestInterface
-    {
-        if (!$this->useCookies) {
-            return $request;
-        }
-
-        foreach ($this->cookieJar->getFor($request->getUri()) as $cookie) {
-            $request = $request->withAddedHeader('Cookie', $cookie->__toString());
-        }
-
-        return $request;
-    }
-
-    /**
-     * @param string[] $headers
-     * @return string[]
-     */
-    protected function sanitizeResponseHeaders(array $headers): array
-    {
-        foreach ($headers as $key => $value) {
-            $headers[$key] = explode(PHP_EOL, $value)[0];
-        }
-
-        return $headers;
-    }
-
-    /**
-     * @param mixed[] $headers
-     * @return array<string, string>
-     */
-    protected function prepareRequestHeadersForHeadlessBrowser(array $headers = []): array
-    {
-        $headers = $this->removeHeadersCausingErrorWithHeadlessBrowser($headers);
-
-        return array_map(function ($headerValue) {
-            return is_array($headerValue) ? reset($headerValue) : $headerValue;
-        }, $headers);
-    }
-
-    /**
-     * @param mixed[] $headers
-     * @return mixed[]
-     */
-    protected function removeHeadersCausingErrorWithHeadlessBrowser(array $headers = []): array
-    {
-        $removeHeaders = ['host'];
-
-        foreach ($headers as $headerName => $headerValue) {
-            if (in_array(strtolower($headerName), $removeHeaders, true)) {
-                unset($headers[$headerName]);
-            }
-        }
-
-        return $headers;
+        return $this->browserHelper()->getBrowser($request);
     }
 }
